fix bug where only the reference to a reverted transaction in block_transactions would be deleted instead of the transaction itself if there were no other references to it
